% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bank}
\alias{bank}
\title{Bank Dataset}
\format{
A data frame with 100 rows and 2 variables:
\describe{
\item{\code{arrival_time}}{The time of each arrival.}
\item{\code{service_time}}{How long each arrival is serviced for.}
}
}
\usage{
bank
}
\description{
Provides simulated data for arrival times and service times.
}
\keyword{datasets}
