% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Multiserver}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{A list of arrival times.}

\item{ServiceTimes}{A list of arrival times.}

\item{NumServers}{A numeric vector.}
}
\value{
Returns a data frame with the simulation.
}
\description{
This function simulates customers going through a queuing system.
}
\examples{
Multiserver(Arrivals=c(5, 10, 15), ServiceTimes=c(5, 15, 30), NumServers=1)
}
